+++ app/Http/Controllers/API/v1/ApiBetController.php
@@ -14,7 +14,7 @@ class ApiBetController extends Controller
 {
     public function index(Request $request, $game, $draw)
     {
-        $request->user()->can('list-bets', Bet::class);
+        $this->authorize('list-bets', Bet::class);
         $game = Game::where('id', $game)
             ->with(['gameConfiguration', 'controlCombination', 'bets'])
             ->get();
@@ -28,13 +28,13 @@ class ApiBetController extends Controller
 
     public function create(Request $request)
     {
-        $request->user()->can('create-bet-transactions', Bet::class);
+        $this->authorize('create-bet-transactions', Bet::class);
         return response([], 200);
     }
 
     public function store(Request $request)
     {
-        $request->user()->can('create-bet-transactions', Bet::class);
+        $this->authorize('create-bet-transactions', Bet::class);
         $validated = $request->validate([
             'agent_id' => 'required',
             'bets.*' => 'required',
@@ -65,19 +65,19 @@ class ApiBetController extends Controller
 
     public function show(Request $request, Bet $bet)
     {
-        $request->user()->can('view-bet-transactions', Bet::class);
+        $this->authorize('view-bet-transactions', Bet::class);
         return response([], 204);
     }
 
     public function edit(Request $request, Bet $bet)
     {
-        $request->user()->can('update-bet-transactions', $bet);
+        $this->authorize('update-bet-transactions', $bet);
         return response(['bets' => $bet], 200);
     }
 
     public function update(Request $request, Bet $bet)
     {
-        $request->user()->can('update-bet-transactions', $bet);
+        $this->authorize('update-bet-transactions', $bet);
         $validated = $request->validated();
         $bet->update($validated);
         return response([$bet], 202);
@@ -85,14 +85,14 @@ class ApiBetController extends Controller
 
     public function destroy(Request $request, Bet $bet)
     {
-        $request->user()->can('delete-bet-transactions', $bet);
+        $this->authorize('delete-bet-transactions', $bet);
         $bet->delete();
         return response([], 204);
     }
 
     public function getBetsRange(Request $request, $game, $date)
     {
-        $request->user()->can('list-bet-transactions', Bet::class);
+        $this->authorize('list-bet-transactions', Bet::class);
         $bets = Bet::withoutGlobalScope(BetScope::class)
             ->where('created_at', 'like', "{$date}%")
             ->with(['betTransaction.user', 'game', 'drawPeriod'])->get();
@@ -106,7 +106,7 @@ class ApiBetController extends Controller
 
     public function getGeneralBetsReport(Request $request)
     {
-        $request->user()->can('list-bet-transactions', Bet::class);
+        $this->authorize('list-bet-transactions', Bet::class);
         $validated = $request->validate([
             'cluster_id' => 'required|array',
             'draw_period_id' => 'required|array',
@@ -130,7 +130,7 @@ class ApiBetController extends Controller
 
     public function getCombinationBetsReport(Request $request)
     {
-        $request->user()->can('list-bet-transactions', Bet::class);
+        $this->authorize('list-bet-transactions', Bet::class);
         $validated = $request->validate([
             'cluster_id' => 'required|array',
             'draw_period_id' => 'required|array',
