+++ app/Http/Controllers/API/v1/ApiControlledNumberController.php
@@ -25,7 +25,7 @@ class ApiControlledNumberController extends Controller
 
     public function store(Request $request, $game)
     {
-        $request->user()->can('create-control-combinations', ControlCombination::class);
+        $this->authorize('create-control-combinations', ControlCombination::class);
         $validated = $request->validate([
             "combination" => "required",
             "max_amount" => "required"
@@ -62,7 +62,7 @@ class ApiControlledNumberController extends Controller
 
     public function update(Request $request, ControlCombination $controlCombination, $game)
     {
-        $request->user()->can('update-control-combinations', $controlCombination);
+        $this->authorize('update-control-combinations', $controlCombination);
         $validated = $request->validate([
             'id' => 'required',
             "combination" => "required",
@@ -85,7 +85,7 @@ class ApiControlledNumberController extends Controller
 
     public function destroy(Request $request, ControlCombination $combi)
     {
-        $request->user()->can('delete-control-combinations', $combi);
+        $this->authorize('delete-control-combinations', $combi);
         $combi->delete();
         broadcast(new NewControlledBetAdded(Game::find($combi->game_id)));
         return response([], 204);
