+++ app/Http/Controllers/API/v1/ApiPermissionController.php
@@ -16,7 +16,7 @@ class ApiPermissionController extends Controller
 
     public function index(Request $request)
     {
-        $request->user()->can('list-permissions', Permission::class);
+        $this->authorize('list-permissions', Permission::class);
         $search = $request->get('search', '');
         $permissions = Permission::with('roles')->get();
         return response(['permissions' => $permissions], 200);
@@ -24,13 +24,13 @@ class ApiPermissionController extends Controller
 
     public function create(Request $request)
     {
-        $request->user()->can('create-permissions', Permission::class);
+        $this->authorize('create-permissions', Permission::class);
         return response([], 200);
     }
 
     public function store(Request $request)
     {
-        $request->user()->can('create-permissions', Permission::class);
+        $this->authorize('create-permissions', Permission::class);
         $validated = $request->validated();
         $permission = Permission::create($validated);
 
@@ -39,19 +39,19 @@ class ApiPermissionController extends Controller
 
     public function show(Request $request, Permission $permission)
     {
-        $request->user()->can('view-permissions', $permission);
+        $this->authorize('view-permissions', $permission);
         return response([], 204);
     }
 
     public function edit(Request $request, Permission $permission)
     {
-        $request->user()->can('update-permissions', $permission);
+        $this->authorize('update-permissions', $permission);
         return response(['permission' => $permission], 200);
     }
 
     public function update(Request $request, Permission $permission)
     {
-        $request->user()->can('update-permissions', $permission);
+        $this->authorize('update-permissions', $permission);
         $validated = $request->validate([
             'roles' => 'array',
         ]);
@@ -65,7 +65,7 @@ class ApiPermissionController extends Controller
 
     public function destroy(Request $request, Permission $permission)
     {
-        $request->user()->can('delete-permissions', $permission);
+        $this->authorize('delete-permissions', $permission);
         $permission->delete();
         return response([], 204);
     }
