+++ app/Http/Controllers/API/v1/ApiUserController.php
@@ -19,7 +19,7 @@ class ApiUserController extends Controller
 {
     public function index(Request $request)
     {
-        $request->user()->can('list-users', User::class);
+        $this->authorize('list-users', User::class);
         $search = $request->get('search', '');
         $users = User::search($search)->with(['cluster', 'address', 'roles'])->get();
 //
@@ -28,7 +28,7 @@ class ApiUserController extends Controller
 
     public function create(Request $request)
     {
-        $request->user()->can('create-users', User::class);
+        $this->authorize('create-users', User::class);
         $clusters = Cluster::pluck('name', 'id');
         $roles = Role::get();
         return response(['clusters' => $clusters, 'roles' => $roles], 200);
@@ -36,7 +36,7 @@ class ApiUserController extends Controller
 
     public function store(Request $request)
     {
-        $request->user()->can('create-users', User::class);
+        $this->authorize('create-users', User::class);
         $validated = $request->validate([
             'role.*' => 'required',
             'name' => 'required',
@@ -74,14 +74,14 @@ class ApiUserController extends Controller
 
     public function show(Request $request, User $user)
     {
-        $request->user()->can('view-users', $user);
+        $this->authorize('view-users', $user);
         return response(['user' => $user], 200);
 //        return view('app.users.show', compact('user'));
     }
 
     public function edit(Request $request, User $user)
     {
-        $request->user()->can('update-users', $user);
+        $this->authorize('update-users', $user);
         $bases = Base::pluck('base_name', 'id');
         $roles = Role::get();
         return response(['bases' => $bases, 'roles' => $roles], 200);
@@ -89,7 +89,7 @@ class ApiUserController extends Controller
 
     public function update(Request $request, User $user)
     {
-        $request->user()->can('update-users', $user);
+        $this->authorize('update-users', $user);
         $validated = $request->validated();
         if (empty($validated['password'])) {
             unset($validated['password']);
@@ -103,7 +103,7 @@ class ApiUserController extends Controller
 
     public function destroy(Request $request, User $user)
     {
-        $request->user()->can('delete-users', $user);
+        $this->authorize('delete-users', $user);
         $user->delete();
         return response([], 204);
     }
@@ -111,7 +111,7 @@ class ApiUserController extends Controller
 
     public function baseRoleIndex(Request $request, $role)
     {
-        $request->user()->can('list-users', User::class);
+        $this->authorize('list-users', User::class);
         $search = $request->get('search', '');
         $users = User::search($search)->with(['cluster', 'address', 'roles' => function ($query) use ($role) {
             $query->whereIn('name', [$role]);
